Validate TP gas flow field against exact Sod on uniform mesh (M1)
Rebuild the multicomponent thermally-perfect flow-field validation on a
uniform 1D shock-tube mesh (case/05.sod_shock_tube, nx=499, all-slip,
inviscid SLAU, MUSCL+limiter, RK3, fixed dt) instead of the earlier
wall-clustered flat_plate mesh whose dt was prohibitively small.

N2 shock tube (PL/PR=10, TL/TR=2000/400K) run to t=2.5e-4s for both
thermalMethod==2 (TP, NASA-9) and thermalMethod==0 (CPG reference):
- Framework check (CPG vs exact Sod Riemann): centerline L2 relative
  error P=0.90%, u=2.96%, rho=3.53% -> TP flux+EOS+BC numerics match
  the exact solution.
- cp(T) effect (TP vs CPG): TP runs hotter in the ~1650K plateau and
  shifts contact/shock fronts faster (T diff up to 248K, u diff 225m/s),
  quantifying the real-gas effect.

Add gen_shocktube_ic.py (TP/CPG IC) and compare_shocktube.py (exact Sod
solver + centerline comparison). Make geometric-metric computation in
calcStructualVariables_d FP64 for double consistency with TP thermo.

+++ solver_density_cuda/cuda_forge/calcStructualVariables_d.cu
@@ -33,17 +33,17 @@ void calcStructualVariables_d
         geom_float dcx = ccx1 - ccx0;
         geom_float dcy = ccy1 - ccy0;
         geom_float dcz = ccz1 - ccz0;
-        geom_float dc  = sqrtf( powf(dcx, 2.0) + pow(dcy, 2.0) + pow(dcz, 2.0));
+        geom_float dc  = sqrt( pow(dcx, 2.0) + pow(dcy, 2.0) + pow(dcz, 2.0));
 
         geom_float dc0px = sx[ip]*(pcx[ip] - ccx0)/ss[ip];
         geom_float dc0py = sy[ip]*(pcy[ip] - ccy0)/ss[ip];
         geom_float dc0pz = sz[ip]*(pcz[ip] - ccz0)/ss[ip];
-        geom_float dc0p  = sqrtf( powf(dc0px, 2.0) + pow(dc0py, 2.0) + pow(dc0pz, 2.0));
+        geom_float dc0p  = sqrt( pow(dc0px, 2.0) + pow(dc0py, 2.0) + pow(dc0pz, 2.0));
 
         geom_float dc1px = sx[ip]*(pcx[ip] - ccx1)/ss[ip];
         geom_float dc1py = sy[ip]*(pcy[ip] - ccy1)/ss[ip];
         geom_float dc1pz = sz[ip]*(pcz[ip] - ccz1)/ss[ip];
-        geom_float dc1p  = sqrtf( powf(dc1px, 2.0) + pow(dc1py, 2.0) + pow(dc1pz, 2.0));
+        geom_float dc1p  = sqrt( pow(dc1px, 2.0) + pow(dc1py, 2.0) + pow(dc1pz, 2.0));
 
         fx [ip] = dc1p/(dc0p + dc1p);
         dcc[ip] = dc;
